Task_1_Python_Introduction/requirements.txt
mysql-connector-python
dict2xml
pre-commit
black
isort
flake8